+++ changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 1.10.0
+- Scenes can listen for a `onDeactivate` event function
+    - can be used for timers, and event unbinding
 ## 1.9.0
 - Improve the Debug Panel
   - Displays the current and previous scene.